fix: rename signin-schema copy to signin-schema
